Implementing working SettingsModalButton
