--- frontend/app-development/features/SettingsModalButton/SettingsModalButton.module.css
@@ -1,3 +1,0 @@
-.button {
-  margin: 100px;
-}
