+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/SettingsModal.test.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
-import { render, screen, act } from '@testing-library/react';
+import {
+  render as rtlRender,
+  screen,
+  act,
+  waitForElementToBeRemoved,
+} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { SettingsModal, SettingsModalProps } from './SettingsModal';
 import { textMock } from '../../../../testing/mocks/i18nMock';
@@ -9,14 +14,52 @@ import { ServicesContextProps, ServicesContextProvider } from 'app-shared/contex
 import { queriesMock } from 'app-shared/mocks/queriesMock';
 import { AppConfig } from 'app-shared/types/AppConfig';
 import { useAppConfigMutation } from 'app-development/hooks/mutations';
+import { mockPolicy } from './mocks/policyMock';
 import { mockAppConfig } from './mocks/appConfigMock';
 import { mockRepository1 } from './mocks/repositoryMock';
-import { mockPolicy } from './mocks/policyMock';
 import { mockAppMetadata } from './mocks/applicationMetadataMock';
+import { Commit, CommitAuthor } from 'app-shared/types/Commit';
+import { MemoryRouter } from 'react-router-dom';
 
 const mockApp: string = 'app';
 const mockOrg: string = 'org';
-const mockCreatedBy: string = 'Mock Mockesen';
+
+const mockCommitAuthor: CommitAuthor = {
+  email: '',
+  name: 'Mock Mockesen',
+  when: new Date(2023, 9, 22),
+};
+
+const mockInitialCommit: Commit = {
+  message: '',
+  author: mockCommitAuthor,
+  comitter: mockCommitAuthor,
+  sha: '',
+  messageShort: '',
+  encoding: '',
+};
+
+const getAppPolicy = jest.fn().mockImplementation(() => Promise.resolve({}));
+const getAppConfig = jest.fn().mockImplementation(() => Promise.resolve({}));
+const getRepoMetadata = jest.fn().mockImplementation(() => Promise.resolve({}));
+const getRepoInitialCommit = jest.fn().mockImplementation(() => Promise.resolve({}));
+const getAppMetadata = jest.fn().mockImplementation(() => Promise.resolve({}));
+
+const resolveMocks = () => {
+  getAppPolicy.mockImplementation(() => Promise.resolve(mockPolicy));
+  getAppConfig.mockImplementation(() => Promise.resolve(mockAppConfig));
+  getRepoMetadata.mockImplementation(() => Promise.resolve(mockRepository1));
+  getRepoInitialCommit.mockImplementation(() => Promise.resolve(mockInitialCommit));
+  getAppMetadata.mockImplementation(() => Promise.resolve(mockAppMetadata));
+};
+
+jest.mock('react-router-dom', () => ({
+  ...jest.requireActual('react-router-dom'),
+  useParams: () => ({
+    app: mockApp,
+    org: mockOrg,
+  }),
+}));
 
 jest.mock('../../../hooks/mutations/useAppConfigMutation');
 const updateAppConfigMutation = jest.fn();
@@ -36,17 +79,64 @@ describe('SettingsModal', () => {
   const defaultProps: SettingsModalProps = {
     isOpen: true,
     onClose: mockOnClose,
-    policy: mockPolicy,
     org: mockOrg,
     app: mockApp,
-    appConfig: mockAppConfig,
-    repository: mockRepository1,
-    createdBy: mockCreatedBy,
-    appMetadata: mockAppMetadata,
   };
 
+  it('fetches policy on mount', () => {
+    render(defaultProps);
+    expect(getAppPolicy).toHaveBeenCalledTimes(1);
+  });
+
+  it('fetches appConfig on mount', () => {
+    render(defaultProps);
+    expect(getAppConfig).toHaveBeenCalledTimes(1);
+  });
+
+  it('fetches repoMetaData on mount', () => {
+    render(defaultProps);
+    expect(getRepoMetadata).toHaveBeenCalledTimes(1);
+  });
+
+  it('fetches commit data on mount', () => {
+    render(defaultProps);
+    expect(getRepoInitialCommit).toHaveBeenCalledTimes(1);
+  });
+
+  it('fetches appMetadata on mount', () => {
+    render(defaultProps);
+    expect(getAppMetadata).toHaveBeenCalledTimes(1);
+  });
+
+  it('initially displays the spinner when loading data', () => {
+    render(defaultProps);
+
+    expect(screen.getByTitle(textMock('settings_modal.loading_content'))).toBeInTheDocument();
+  });
+
+  it.each([
+    'getAppPolicy',
+    'getAppConfig',
+    'getAppMetadata',
+    'getRepoMetadata',
+    'getRepoInitialCommit',
+  ])('shows an error message if an error occured on the %s query', async (queryName) => {
+    const errorMessage = 'error-message-test';
+    render(defaultProps, {
+      [queryName]: () => Promise.reject({ message: errorMessage }),
+    });
+
+    await waitForElementToBeRemoved(() =>
+      screen.queryByTitle(textMock('settings_modal.loading_content')),
+    );
+
+    expect(screen.getByText(textMock('general.fetch_error_message'))).toBeInTheDocument();
+    expect(screen.getByText(textMock('general.error_message_with_colon'))).toBeInTheDocument();
+    expect(screen.getByText(errorMessage)).toBeInTheDocument();
+  });
+
   it('closes the modal when the close button is clicked', async () => {
-    render(<SettingsModal {...defaultProps} isOpen />);
+    render(defaultProps);
 
     const closeButton = screen.getByRole('button', { name: textMock('modal.close_icon') });
     await act(() => user.click(closeButton));
@@ -54,8 +144,9 @@ describe('SettingsModal', () => {
     expect(mockOnClose).toHaveBeenCalledTimes(1);
   });
 
-  it('displays left navigation bar on mount', () => {
-    render(<SettingsModal {...defaultProps} />);
+  it('displays left navigation bar when promises resolves', async () => {
+    await resolveAndWaitForSpinnerToRemove();
+
     expect(
       screen.getByRole('button', { name: textMock('settings_modal.left_nav_tab_about') }),
     ).toBeInTheDocument();
@@ -70,8 +161,8 @@ describe('SettingsModal', () => {
     ).toBeInTheDocument();
   });
 
-  it('displays the about tab, and not the other tabs, when modal opens first time', () => {
-    render(<SettingsModal {...defaultProps} />);
+  it('displays the about tab, and not the other tabs, when promises resolves first time', async () => {
+    await resolveAndWaitForSpinnerToRemove();
 
     expect(screen.getByText(textMock('settings_modal.about_tab_heading'))).toBeInTheDocument();
     expect(
@@ -89,7 +180,7 @@ describe('SettingsModal', () => {
   });
 
   it('changes the tab from "about" to "policy" when policy tab is clicked', async () => {
-    renderWithQueryClient({}, createQueryClientMock(), defaultProps);
+    await resolveAndWaitForSpinnerToRemove();
 
     expect(
       screen.queryByRole('heading', {
@@ -116,7 +207,7 @@ describe('SettingsModal', () => {
   });
 
   it('changes the tab from "policy" to "about" when about tab is clicked', async () => {
-    renderWithQueryClient({}, createQueryClientMock(), defaultProps);
+    await resolveAndWaitForSpinnerToRemove();
 
     const policyTab = screen.getByRole('button', {
       name: textMock('settings_modal.left_nav_tab_policy'),
@@ -138,7 +229,7 @@ describe('SettingsModal', () => {
   });
 
   it('changes the tab from "about" to "localChanges" when local changes tab is clicked', async () => {
-    renderWithQueryClient({}, createQueryClientMock(), defaultProps);
+    await resolveAndWaitForSpinnerToRemove();
 
     expect(
       screen.queryByRole('heading', {
@@ -164,7 +255,7 @@ describe('SettingsModal', () => {
   });
 
   it('changes the tab from "about" to "accessControl" when access control tab is clicked', async () => {
-    renderWithQueryClient({}, createQueryClientMock(), defaultProps);
+    await resolveAndWaitForSpinnerToRemove();
 
     expect(
       screen.queryByRole('heading', {
@@ -189,21 +280,40 @@ describe('SettingsModal', () => {
       screen.queryByText(textMock('settings_modal.about_tab_heading')),
     ).not.toBeInTheDocument();
   });
+
+  /**
+   * Resolves the mocks, renders the component and waits for the spinner
+   * to be removed from the screen
+   */
+  const resolveAndWaitForSpinnerToRemove = async () => {
+    resolveMocks();
+    render(defaultProps);
+
+    await waitForElementToBeRemoved(() =>
+      screen.queryByTitle(textMock('settings_modal.loading_content')),
+    );
+  };
 });
 
-const renderWithQueryClient = (
+const render = (
+  props: SettingsModalProps,
   queries: Partial<ServicesContextProps> = {},
   queryClient: QueryClient = createQueryClientMock(),
-  props: SettingsModalProps,
 ) => {
   const allQueries: ServicesContextProps = {
     ...queriesMock,
+    getAppPolicy,
+    getAppConfig,
+    getRepoMetadata,
+    getRepoInitialCommit,
+    getAppMetadata,
     ...queries,
   };
-
-  return render(
-    <ServicesContextProvider {...allQueries} client={queryClient}>
-      <SettingsModal {...props} />
-    </ServicesContextProvider>,
+  return rtlRender(
+    <MemoryRouter>
+      <ServicesContextProvider {...allQueries} client={queryClient}>
+        <SettingsModal {...props} />
+      </ServicesContextProvider>
+    </MemoryRouter>,
   );
 };
