+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/mocks/appConfigMock.ts
