+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/mocks/applicationMetadataMock.ts
