+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/mocks/policyMock.ts
