+++ frontend/app-development/layout/SettingsModalButton/SettingsModal/utils/settingsModalUtils.ts
