+++ frontend/packages/shared/src/components/GiteaHeader/GiteaHeader.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import classes from './GiteaHeader.module.css';
 import { VersionControlButtons } from './VersionControlButtons';
 import { ThreeDotsMenu } from './ThreeDotsMenu';
@@ -21,9 +21,13 @@ type GiteaHeaderProps = {
    */
   hasCloneModal?: boolean;
   /**
-   * Classname for some extra styling
+   * Classname for some extra styling for the right content
    */
-  className?: string;
+  rightContentClassName?: string;
+  /**
+   * Component to be shown on the left
+   */
+  leftComponent?: ReactNode;
 };
 
 /**
@@ -34,14 +38,15 @@ type GiteaHeaderProps = {
  *        app={app}
  *        menuOnlyHasRepository
  *        extraPadding
- *        className={classes.someExtraStyle}
+ *        rightContentClassName={classes.someExtraStyle}
  *    />
  *
  * @property {string}[org] - The name of the organisation
  * @property {string}[app] - The name of the app / repository
  * @property {boolean}[menuOnlyHasRepository] - Flag for if the three dots menu only should show the repository option. This is relevant for resourceadm
  * @property {boolean}[hasCloneModal] - Flag for if the component has a clone modal. This is relevant for app-development
- * @property {string}[className] - Classname for some extra styling
+ * @property {string}[rightContentClassName] - Classname for some extra styling for the right content
+ * @property {ReactNode}[leftComponent] - Component to be shown on the left
  *
  * @returns {React.ReactNode} - The rendered Gitea header component
  */
@@ -50,16 +55,15 @@ export const GiteaHeader = ({
   app,
   menuOnlyHasRepository = false,
   hasCloneModal = false,
-  className,
+  rightContentClassName,
+  leftComponent,
 }: GiteaHeaderProps): React.ReactNode => {
   return (
     <div className={classes.wrapper}>
-      <div className={`${classes.contentWrapper} ${className && className}`}>
+      <div className={classes.leftContentWrapper}>{leftComponent}</div>
+      <div className={`${classes.rightContentWrapper} ${rightContentClassName}`}>
         <VersionControlButtons org={org} app={app} />
-        <ThreeDotsMenu
-          onlyShowRepository={menuOnlyHasRepository}
-          hasCloneModal={hasCloneModal}
-        />
+        <ThreeDotsMenu onlyShowRepository={menuOnlyHasRepository} hasCloneModal={hasCloneModal} />
       </div>
     </div>
   );
